+++ SuperNewRoles/AllRoleSetClass.cs
@@ -5,7 +5,7 @@ using Hazel;
 
 
 using SuperNewRoles.Helpers;
-using SuperNewRoles.Intro;
+;
 using SuperNewRoles.Mode;
 using SuperNewRoles.Mode.SuperHostRoles;
