+++ SuperNewRoles/Mode/SuperHostRoles/RoleChat.cs
@@ -1,6 +1,6 @@
 using System.Collections.Generic;
 using InnerNet;
-using SuperNewRoles.Intro;
+;
 
 namespace SuperNewRoles.Mode.SuperHostRoles
 {
