+++ SuperNewRoles/Mode/SuperHostRoles/RoleSelectHandler.cs
@@ -1,7 +1,7 @@
 using System.Collections.Generic;
 
 
-using SuperNewRoles.Intro;
+;
 using SuperNewRoles.Roles;
 
 namespace SuperNewRoles.Mode.SuperHostRoles
