+++ SuperNewRoles/Mode/Werewolf/main.cs
@@ -1,6 +1,6 @@
 using System.Collections.Generic;
 using SuperNewRoles.Helpers;
-using SuperNewRoles.Intro;
+;
 using SuperNewRoles.Mode.SuperHostRoles;
 using SuperNewRoles.Roles;
