+++ SuperNewRoles/Modules/IntroDate.cs
@@ -6,8 +6,24 @@ using SuperNewRoles.Patch;
 using SuperNewRoles.Roles;
 using UnityEngine;
 
-namespace SuperNewRoles.Intro
+namespace SuperNewRoles.Modules
 {
+    public class TeamDate
+    {
+        public string NameKey;
+        public Color color;
+        public Color BackGround;
+        public List<RoleId> RoleIds;
+
+        TeamDate(string NameKey, Color color, Color BackGround, List<RoleId> RoleId)
+        {
+            this.color = color;
+            this.BackGround = BackGround;
+            this.NameKey = NameKey;
+            RoleIds = RoleId;
+        }
+        public static TeamDate VultureTeam = new("Test", Color.black, Color.yellow, new List<RoleId> { RoleId.Sheriff });
+    }
     public class IntroDate
     {
         public static List<IntroDate> IntroDatas = new();
