+++ SuperNewRoles/Patch/ChatHandler.cs
@@ -5,7 +5,7 @@ using System.Linq;
 using BepInEx.IL2CPP.Utils;
 using HarmonyLib;
 
-using SuperNewRoles.Intro;
+;
 using SuperNewRoles.Mode.SuperHostRoles;
 using SuperNewRoles.Roles;
 using UnityEngine;
