+++ SuperNewRoles/Patch/CustomOptionModel.cs
@@ -7,7 +7,7 @@ using BepInEx.Configuration;
 using HarmonyLib;
 using Hazel;
 
-using SuperNewRoles.Intro;
+;
 using SuperNewRoles.Mode;
 using UnityEngine;
 using UnityEngine.Events;
