+++ SuperNewRoles/Patch/PlayerControlPatch.cs
@@ -8,7 +8,7 @@ using InnerNet;
 
 
 using SuperNewRoles.Helpers;
-using SuperNewRoles.Intro;
+;
 using SuperNewRoles.Mode;
 using SuperNewRoles.Mode.SuperHostRoles;
 using SuperNewRoles.Patch;
