+++ SuperNewRoles/Patch/SetNames.cs
@@ -3,7 +3,7 @@ using System.Collections.Generic;
 using System.Linq;
 
 
-using SuperNewRoles.Intro;
+;
 using SuperNewRoles.Mode;
 using SuperNewRoles.Roles;
 using TMPro;
