+++ SuperNewRoles/PlayerControlHepler.cs
@@ -2,7 +2,7 @@ using System.Collections.Generic;
 using System.Linq;
 using InnerNet;
 
-using SuperNewRoles.Intro;
+;
 using SuperNewRoles.Roles;
 using UnityEngine;
 using static SuperNewRoles.Patch.ShareGameVersion;
