+++ SuperNewRoles/Roles/Ghost/HandleGhostRole.cs
@@ -2,7 +2,7 @@ using System.Collections.Generic;
 using System.Linq;
 using HarmonyLib;
 
-using SuperNewRoles.Intro;
+;
 using SuperNewRoles.Mode;
 
 namespace SuperNewRoles.Roles
