+++ SuperNewRoles/Roles/Impostor/Assassin.cs
@@ -4,7 +4,7 @@ using HarmonyLib;
 
 
 using SuperNewRoles.Helpers;
-using SuperNewRoles.Intro;
+;
 using SuperNewRoles.Mode;
 using SuperNewRoles.Mode.SuperHostRoles;
 using UnhollowerBaseLib;
