+++ SuperNewRoles/Roles/Role/RoleHelper.cs
@@ -3,7 +3,7 @@ using System.Collections.Generic;
 using Hazel;
 using SuperNewRoles.CustomObject;
 
-using SuperNewRoles.Intro;
+;
 using SuperNewRoles.Mode;
 using SuperNewRoles.Roles;
